remove console logs and reorganize constants
