+++ src/Kanbas/Courses/Quizzes/Preview/index.tsx
@@ -1,18 +1,12 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import QuizHeaderPreview from './QuizPreviewHeader';
 import QuizContent from './QuizContent';
 import QuestionList from './QuestionList';
-import { quizzes } from '../../../Database';
-import { useNavigate, useParams } from 'react-router';
+import { useParams } from 'react-router';
 import { QuizPreviewProvider } from '../context/QuizPreviewContext';
 import * as client from './client';
 import StatusBanner from './StatusBanner';
-
-export enum QuestionType {
-  MULTIPLE_CHOICE = 'Multiple Choice',
-  TRUE_FALSE = 'True/False',
-  FILL_IN_THE_BLANK = 'Fill in the Blank',
-}
+import { QuestionType, QuizType, AssignmentGroup } from './constants';
 
 export type Question = {
   _id: number;
@@ -23,20 +17,6 @@ export type Question = {
   answers: string[];
   options: string[];
 };
-
-export enum QuizType {
-  GRADED_QUIZ = 'Graded Quiz',
-  PRACTICE_QUIZ = 'Practice Quiz',
-  GRADED_SURVEY = 'Graded Survey',
-  UNGRADED_SURVEY = 'Ungraded Survey',
-}
-
-export enum AssignmentGroup {
-  QUIZZES = 'Quizzes',
-  ASSIGNMENTS = 'Assignments',
-  EXAMS = 'Exams',
-  PROJECTS = 'Projects',
-}
 
 export type Quiz = {
   _id: number;
@@ -60,28 +40,10 @@ export type Quiz = {
   untilDate: Date;
 };
 
-/*
-
-preview needs:
-- quiz title
-banner to say that it's a preview
-- date quiz started at (assume it's the datetime the moment the faculty click to preview the quiz)
-- show question either one at a time or all at once in a long vertical scroll
-- each question needs to show:
-  - title in top left
-  - description
-  - points in top right
-  - list of all answers in the proper input format
-  - next button to go to next question if there are more questions (dont show if it's the last question)
-- show submit button for overall quiz below question box and with the last updated at time next to it
-- show button below submit for going back to quiz editor screen for the particular question
-- at bottom, show list of all questions where the links are all clickable and the current question link is bolded
-*/
 const QuizPreview = () => {
   const { quizId } = useParams();
 
   const [quiz, setQuiz] = useState<Quiz | null>(null);
-  // TODO: the datatype of this array might need to change
   const [answers, setAnswers] = useState<string[][]>([]);
   const [taggedQuestions, setTaggedQuestions] = useState<number[]>([]);
   const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0);
@@ -91,7 +53,6 @@ const QuizPreview = () => {
   useEffect(() => {
     // fetch quiz from BE
     const fetchQuiz = async () => {
-      // TODO: replace with actual fetch call
       if (!quizId) {
         return;
       }
@@ -119,17 +80,12 @@ const QuizPreview = () => {
     setCurrentQuestionIndex(index);
 
   const handleSubmit = () => {
-    console.log(
-      'grade teh quiz score for the preview test run but dont save it anywhere?'
-    );
-    console.log('submitted answers', answers);
     const results = answers.map(
       (userAnswers, index) =>
         quiz?.questions[index].answers.every(
           (ans, i) => ans === userAnswers[i]
         ) ?? false
     );
-    console.log('results', results);
     setQuizResults(results);
   };
 
@@ -138,9 +94,6 @@ const QuizPreview = () => {
     answerIndex: number,
     questionIndex: number
   ) => {
-    console.log(
-      'save the progress of the quiz after each answer change in browser state'
-    );
     setAnswers((prevAnswers) => {
       const newAnswers = [...prevAnswers];
       newAnswers[questionIndex][answerIndex] = answer;
